requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiogram==2.25.1
aiohttp[speedups]==3.8.4
autopep8==2.0.1
ipykernel==6.21.1
langchain==0.0.150
openai==0.27.4
pre-commit==3.0.4
python_dotenv==1.0.0
pymilvus==2.2.7
seaborn==0.12.2
tiktoken==0.3.3
transformers==4.28.1
torch
torchvision
ujson==5.7.0
uvloop==0.17.0
